Extract Excel workbook selection helpers
